Extract BaseAPI superclass
