+++ kentik_api_library/kentik_api/api_resources/users_api.py
@@ -2,26 +2,22 @@ from http import HTTPStatus
 from typing import List
 
 from kentik_api.api_calls import users
-from kentik_api.api_connection.api_connector_protocol import APIConnectorProtocol
+from kentik_api.api_resources.base_api import BaseAPI
 from kentik_api.public.user import User
 from kentik_api.requests_payload import users_payload
-from kentik_api.requests_payload.as_dict import as_dict
 
 
-class UsersAPI:
+class UsersAPI(BaseAPI):
     """Exposes Kentik API operations related to users. """
-
-    def __init__(self, api_connector: APIConnectorProtocol) -> None:
-        self._api_connector = api_connector
 
     def get_all(self) -> List[User]:
         api_call = users.get_users()
-        response = self._api_connector.send(api_call)
+        response = self._send(api_call)
         return users_payload.GetAllResponse.from_json(response.text).to_users()
 
     def get(self, user_id: int) -> User:
         api_call = users.get_user_info(user_id)
-        response = self._api_connector.send(api_call)
+        response = self._send(api_call)
         return users_payload.GetResponse.from_json(response.text).to_user()
 
     def create(self, user: User) -> User:
@@ -32,38 +28,34 @@ class UsersAPI:
         assert user.email_product is not None, "Email product has to be provided"
 
         api_call = users.create_user()
-        payload = as_dict(
-            users_payload.CreateRequest(
-                user_name=user.username,
-                user_full_name=user.full_name,
-                user_email=user.email,
-                user_password=user.password,
-                role=user.role,
-                email_service=user.email_service,
-                email_product=user.email_product,
-            )
+        payload = users_payload.CreateRequest(
+            user_name=user.username,
+            user_full_name=user.full_name,
+            user_email=user.email,
+            user_password=user.password,
+            role=user.role,
+            email_service=user.email_service,
+            email_product=user.email_product,
         )
-        response = self._api_connector.send(api_call, payload)
+        response = self._send(api_call, payload)
         return users_payload.CreateResponse.from_json(response.text).to_user()
 
     def update(self, user: User) -> User:
         assert user.id is not None, "User ID has to be provided"
 
         api_call = users.update_user(user.id)
-        payload = as_dict(
-            users_payload.UpdateRequest(
-                user_name=user.username,
-                user_full_name=user.full_name,
-                user_email=user.email,
-                role=user.role,
-                email_service=user.email_service,
-                email_product=user.email_product,
-            )
+        payload = users_payload.UpdateRequest(
+            user_name=user.username,
+            user_full_name=user.full_name,
+            user_email=user.email,
+            role=user.role,
+            email_service=user.email_service,
+            email_product=user.email_product,
         )
-        response = self._api_connector.send(api_call, payload)
+        response = self._send(api_call, payload)
         return users_payload.UpdateResponse.from_json(response.text).to_user()
 
     def delete(self, user_id: int) -> bool:
         api_call = users.delete_user(user_id)
-        response = self._api_connector.send(api_call)
+        response = self._send(api_call)
         return response.http_status_code == HTTPStatus.NO_CONTENT
